tests/wpt/web-platform-tests/IndexedDB/idbdatabase_deleteObjectStore3.htm

<!DOCTYPE html>
<title>IDBDatabase.deleteObjectStore() - attempt to remove an object store that does not exist </title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>

var t = async_test(),
    open_rq = createdb(t);

open_rq.onupgradeneeded = function(e)
{
    var db = e.target.result;
    assert_throws('NotFoundError',
        function() { db.deleteObjectStore('whatever'); });
    t.done();
}

</script>

<div id="log"></div>
